Working on renaming
Got most of the way through the full page app - still have to update all
of the CSS class names and make sure all string refs in the code update
(webstorm's refactoring was struggling with this :[ ).  And, all of the
examples have to be updated - they are all broken and say vA3C all over
them.